グループ退会時のいくつかの不具合を修正し、次の管理者選択モーダルで次の管理者を選択しなかった場合、最終確認モーダルを表示する機能を実装しました。

+++ app/Http/Livewire/QuitGroupForm.php
@@ -13,30 +13,33 @@ use Illuminate\Support\Facades\Gate;
 
 class QuitGroupForm extends Component
 {
-    public $user_id;
+    public $selectedUserId;
     public $password;
     public $group_data;
     public $showModal = false;
     public $showNextManagerModal = false;
+    public $showModalNobodySubManager = false;
+    public $showModalFinalConfirmation = false;
 
 
     protected $rules = [
         'password' => ['required'],
     ];
 
+
     public function mount()
     {
         $this->group_data = Group::with(['userRoles' => function ($query) {
             $query->wherePivot('role', 50);
         }])->find(session()->get('group_id'));
-
-        // dd($this->group_data);
     }
+
 
     public function getListeners()
     {
         return ['showModal' => 'showModal'];
     }
+
 
     public function updatedShowNextManagerModal($value)
     {
@@ -45,18 +48,24 @@ class QuitGroupForm extends Component
         }
     }
 
+
     public function showModal()
     {
         $this->showModal = true;
     }
 
+
     public function closeModal()
     {
+        $this->selectedUserId = '';
         $this->showModal = false;
         $this->showNextManagerModal = false;
+        $this->showModalNobodySubManager = false;
+        $this->showModalFinalConfirmation = false;
         $this->password = '';
         $this->resetErrorBag();
     }
+
 
     public function quitGroup()
     {
@@ -67,22 +76,33 @@ class QuitGroupForm extends Component
             return;
         }
 
+        $this->password = "";
+
         // パスワードが一致したときの処理
-        //管理者権限の場合
         if (Gate::allows('manager', $this->group_data)) {
-            $this->showModal = false;
-            $this->showNextManagerModal = true;
+            //管理者権限の場合
 
-
+            // グループのデータ（サブ管理者のデータも併せて取得）
             $this->group_data = Group::with(['userRoles' => function ($query) {
                 $query->wherePivot('role', 50);
             }])->find(session()->get('group_id'));
 
-            // dd($this->group_data);
 
+            if ($this->group_data->userRoles->isNotEmpty()) {
+                //サブ管理者がいる場合                
 
-            // dd($this->showNextManagerModal);
+                // →「サブ管理者から次の管理者を選択するためのモーダル」を表示する
+                $this->showModal = false;
+                $this->showNextManagerModal = true;
+            } else {
+                //サブ管理者がいない場合
+
+                // →「サブ管理者いませんよモーダル」を表示する
+                $this->showModal = false;
+                $this->showModalNobodySubManager = true;
+            }
         } else {
+            // 管理者以外の権限の場合
             $group_data = Group::find(session()->get('group_id'));
             $group_data->user()->detach(Auth::user());
             $group_data->userRoles()->detach(Auth::user());
@@ -94,14 +114,59 @@ class QuitGroupForm extends Component
 
     public function quitGroupForManager()
     {
-        $this->group_data->userRoles()->updateExistingPivot($this->user_id, ['role' => 10]);
+        if ($this->selectedUserId) {
+            // サブ管理者を選択する場合
+
+            // →選択したサブ管理者を管理者にして自分はグループを退会する
+            $this->group_data->userRoles()->updateExistingPivot($this->selectedUserId, ['role' => 10]);
+
+            $this->group_data->user()->detach(Auth::user());
+            $this->group_data->userRoles()->detach(Auth::user());
+
+            return to_route('index');
+        } else {
+            // サブ管理者を選択しない場合
+
+            $this->showNextManagerModal = false;
+            $this->showModalFinalConfirmation = true;
+        }
+    }
 
 
-        $this->group_data->user()->detach(Auth::user());
-        $this->group_data->userRoles()->detach(Auth::user());
+    public function deleteGroup()
+    {
+        $this->validate();
+
+        if (!Hash::check($this->password, Auth::user()->password)) {
+            $this->addError('password', 'パスワードが一致しません。');
+            return;
+        }
+
+        // →グループを削除する
+        $group_data = Group::find(session()->get('group_id'));
+        $group_data->delete();
 
         return to_route('index');
     }
+
+
+    public function quitGroupWhenNobodySubManager()
+    {
+        $this->validate();
+
+        if (!Hash::check($this->password, Auth::user()->password)) {
+            $this->addError('password', 'パスワードが一致しません。');
+            return;
+        }
+
+
+        // →グループを削除する
+        $group_data = Group::find(session()->get('group_id'));
+        $group_data->delete();
+
+        return to_route('index');
+    }
+
 
     public function render()
     {
